split upload tests to a seperate test file
